improved debug possibilities for Android and fixed crash in Android Version
The crash came from using env instead of _env. This was not reentrant
and is e.g. a setmouse from the other thread interrupted the main
process while inside a JNICALL which needs env it could change the env
in the middle of the JNIcall on some devices (Toshiba TABLET was always
OK!). This caused a SIGABRT crash maybe coming from the JNI-Framework
from an assert() or so.

kleinere kosmetische Korrekturen + CHECKENV()

Also: kleine Optimierungen an Eventqueue (framebuffer.c)

release --> 37